+++ tests/xarray/test_samples.py
@@ -17,7 +17,7 @@ import xarray as xr
 from anemoi.datasets.create.functions.sources.xarray import XarrayFieldList
 from anemoi.datasets.testing import assert_field_list
 
-URL = "https://object-store.os-api.cci1.ecmwf.int/ml-tests/test-data/samples/"
+URL = "https://object-store.os-api.cci1.ecmwf.int/ml-tests/test-data/samples/anemoi-datasets/"
 
 SAMPLES = list(range(23))
 SKIP = [0, 1, 2, 3, 4, 22]