+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Keep it human-readable, your future self will thank you!
 - Use cKDTree instead of KDTree
 - Implement 'complement' feature
 - Add ability to patch xarrays (#160)
+- Add support of ECCC file formats (fstd)
+- Add `use_cdsapi_dataset` option to mars and accumulations
 
 ### Added